+++ app/Routes.js
@@ -3,7 +3,7 @@ import {connect} from 'react-redux';
 import {IndexRoute, Router, Route} from 'react-router';
 import ContainerPage from './containers/pages/ContainerPage';
 import HomePage from './containers/pages/HomePage';
-import ScenariosPage from './components/pages/ScenariosPage';
+import MapsPage from './components/pages/MapsPage';
 import CountriesPage from './components/pages/CountriesPage';
 import ComparePage from './components/pages/ComparePage';
 
@@ -12,7 +12,7 @@ const Routes = ({history}) => {
     <Router history={history}>
       <Route path="/" component={ContainerPage}>
         <IndexRoute component={HomePage}/>
-        <Route path="scenarios" component={ScenariosPage}/>
+        <Route path="global-scenarios" component={MapsPage}/>
         <Route path="countries" component={CountriesPage}/>
         <Route path="compare" component={ComparePage}/>
       </Route>
